Remove compilation warnings. Cast overly-specific #def'd values to the destination type, unsigned int.

+++ driver-bflsc.c
@@ -839,7 +839,7 @@ reinit:
 
 	sc_info->scan_sleep_time = BAS_SCAN_TIME;
 	sc_info->results_sleep_time = BFLSC_RES_TIME;
-	sc_info->default_ms_work = BAS_WORK_TIME;
+	sc_info->default_ms_work = (unsigned int)BAS_WORK_TIME;
 	latency = BAS_LATENCY;
 
 	/* When getinfo() "FREQUENCY: [UNKNOWN]" is fixed -
@@ -849,20 +849,20 @@ reinit:
 	if (sc_info->sc_count > 1) {
 		newname = BFLSC_MINIRIG;
 		sc_info->scan_sleep_time = BAM_SCAN_TIME;
-		sc_info->default_ms_work = BAM_WORK_TIME;
+		sc_info->default_ms_work = (unsigned int)BAM_WORK_TIME;
 		bflsc->usbdev->ident = IDENT_BAM;
 		latency = BAM_LATENCY;
 	} else {
 		if (sc_info->sc_devs[0].engines < 34) { // 16 * 2 + 2
 			newname = BFLSC_JALAPENO;
 			sc_info->scan_sleep_time = BAJ_SCAN_TIME;
-			sc_info->default_ms_work = BAJ_WORK_TIME;
+			sc_info->default_ms_work = (unsigned int)BAJ_WORK_TIME;
 			bflsc->usbdev->ident = IDENT_BAJ;
 			latency = BAJ_LATENCY;
 		} else if (sc_info->sc_devs[0].engines < 130)  { // 16 * 8 + 2
 			newname = BFLSC_LITTLESINGLE;
 			sc_info->scan_sleep_time = BAL_SCAN_TIME;
-			sc_info->default_ms_work = BAL_WORK_TIME;
+			sc_info->default_ms_work = (unsigned int)BAL_WORK_TIME;
 			bflsc->usbdev->ident = IDENT_BAL;
 			latency = BAL_LATENCY;
 		}
@@ -872,7 +872,7 @@ reinit:
 	if (sc_info->ident == IDENT_BMA) {
 		bflsc->drv->queue_full = &bflsc28_queue_full;
 		sc_info->scan_sleep_time = BMA_SCAN_TIME;
-		sc_info->default_ms_work = BMA_WORK_TIME;
+		sc_info->default_ms_work = (unsigned int)BMA_WORK_TIME;
 		sc_info->results_sleep_time = BMA_RES_TIME;
 	}
